% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflogix-build.R
\name{leaflogix-build}
\alias{leaflogix-build}
\alias{bld_ll_population}
\alias{bld_ll_transactions}
\alias{bld_ll_stocksnapshot}
\alias{bld_ll_inventory}
\alias{bld_ll_products}
\alias{bld_ll_customers}
\alias{bld_ll_loyalty}
\alias{bld_ll_employees}
\alias{bld_ll_categories}
\alias{bld_ll_brands}
\alias{..bld_ll_customer_summary}
\alias{..bld_ll_store_employees}
\alias{..bld_ll_order_summary}
\alias{..bld_ll_order_items}
\alias{..bld_ll_customer_loyalty}
\alias{..bld_ll_product_summary}
\alias{..bld_ll_stock_lab_results}
\alias{..bld_ll_stock_by_room}
\alias{..bld_ll_stock_summary}
\alias{..bld_ll_stock_snapshot}
\alias{..bld_ll_brand_index}
\alias{..bld_ll_category_index}
\title{Build leaflogix consolidated}
\usage{
bld_ll_population(index, org, store)

bld_ll_transactions(index, org, store)

bld_ll_stocksnapshot(index, org, store)

bld_ll_inventory(index, org, store)

bld_ll_products(index, org, store)

bld_ll_customers(index, org, store)

bld_ll_loyalty(index, org, store)

bld_ll_employees(index, org, store)

bld_ll_categories(index, org, store)

bld_ll_brands(index, org, store)

..bld_ll_customer_summary(index, org = NULL, store = NULL)

..bld_ll_store_employees(index, org = NULL, store = NULL)

..bld_ll_order_summary(index, org = NULL, store = NULL)

..bld_ll_order_items(index, org = NULL, store = NULL)

..bld_ll_customer_loyalty(index, org = NULL, store = NULL)

..bld_ll_product_summary(index, org = NULL, store = NULL)

..bld_ll_stock_lab_results(index, org = NULL, store = NULL)

..bld_ll_stock_by_room(index, org = NULL, store = NULL)

..bld_ll_stock_summary(index, org = NULL, store = NULL)

..bld_ll_stock_snapshot(index, org = NULL, store = NULL)

..bld_ll_brand_index(index, org = NULL, store = NULL)

..bld_ll_category_index(index, org = NULL, store = NULL)
}
\arguments{
\item{index}{output list from extracted step}

\item{org}{optional org for logging}

\item{store}{optional store for logging}
}
\value{
Returns a consolidated table for each endpoint that contains all leaflogix locations
}
\description{
Build leaflogix consolidated
}
\section{Functions}{
\itemize{
\item \code{bld_ll_population()}: build population

\item \code{bld_ll_transactions()}: transactions

\item \code{bld_ll_stocksnapshot()}: stock snapshot

\item \code{bld_ll_inventory()}: inventory

\item \code{bld_ll_products()}: products

\item \code{bld_ll_customers()}: customers

\item \code{bld_ll_loyalty()}: loyalty

\item \code{bld_ll_employees()}: employees

\item \code{bld_ll_categories()}: categories

\item \code{bld_ll_brands()}: brands

\item \code{..bld_ll_customer_summary()}: build consolidated customers

\item \code{..bld_ll_store_employees()}: build consolidated customers

\item \code{..bld_ll_order_summary()}: build consolidated orders table

\item \code{..bld_ll_order_items()}: build consolidated order lines table

\item \code{..bld_ll_customer_loyalty()}: build consolidated loyalty table

\item \code{..bld_ll_product_summary()}: build consolidated products table

\item \code{..bld_ll_stock_lab_results()}: build consolidated inventory tables

\item \code{..bld_ll_stock_by_room()}: build consolidated inventory tables

\item \code{..bld_ll_stock_summary()}: build consolidated inventory tables

\item \code{..bld_ll_stock_snapshot()}: build consolidated inventory tables

\item \code{..bld_ll_brand_index()}: build consolidated brands table

\item \code{..bld_ll_category_index()}: build consolidated brands table

}}
