% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflogix-extract.R
\name{leaflogix-extract}
\alias{leaflogix-extract}
\alias{ext_ll_population}
\alias{ext_ll_inventory}
\alias{ext_ll_brands}
\alias{ext_ll_categories}
\alias{ext_ll_loyalty}
\alias{ext_ll_employees}
\alias{ext_ll_products}
\alias{ext_ll_transactions}
\alias{ext_ll_customers}
\alias{ext_ll_stocksnapshot}
\title{Extract Leaflogix JSONs}
\usage{
ext_ll_population(json, org, store)

ext_ll_inventory(json, org, store)

ext_ll_brands(json, org, store)

ext_ll_categories(json, org, store)

ext_ll_loyalty(json, org, store)

ext_ll_employees(json, org, store)

ext_ll_products(json, org, store)

ext_ll_transactions(json, org, store)

ext_ll_customers(json, org, store)

ext_ll_stocksnapshot(json, org, store)
}
\arguments{
\item{json}{json returned from GET step}

\item{org}{org short name}

\item{store}{store short name}
}
\value{
Extract functions return a data.table containing extracted
jsons except for Transactions, which returns a list of 4 data.tables
}
\description{
Extract Leaflogix JSONs
}
\section{Functions}{
\itemize{
\item \code{ext_ll_population()}: extract population input tables

\item \code{ext_ll_inventory()}: extract inventory

\item \code{ext_ll_brands()}: extract brands

\item \code{ext_ll_categories()}: extract categories

\item \code{ext_ll_loyalty()}: extract loyalty

\item \code{ext_ll_employees()}: extract employees

\item \code{ext_ll_products()}: extract products

\item \code{ext_ll_transactions()}: extract transactions

\item \code{ext_ll_customers()}: extract customers

\item \code{ext_ll_stocksnapshot()}: extract stock_snapshot

}}
