% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflogix-get.R
\name{leaflogix-get}
\alias{leaflogix-get}
\alias{get_ll_population}
\alias{get_ll_inventory}
\alias{get_ll_stocksnapshot}
\alias{get_ll_brands}
\alias{get_ll_categories}
\alias{get_ll_transactions}
\alias{get_ll_products}
\alias{get_ll_loyalty}
\alias{get_ll_employees}
\alias{get_ll_customers}
\title{Leaflogix Get Endpoint Functions}
\usage{
get_ll_population(org, store, auth, consumerkey, ...)

get_ll_inventory(org, store, auth, consumerkey)

get_ll_stocksnapshot(org, store, auth, consumerkey, ...)

get_ll_brands(org, store, auth, consumerkey)

get_ll_categories(org, store, auth, consumerkey)

get_ll_transactions(org, store, auth, consumerkey, ...)

get_ll_products(org, store, auth, consumerkey, ...)

get_ll_loyalty(org, store, auth, consumerkey)

get_ll_employees(org, store, auth, consumerkey)

get_ll_customers(org, store, auth, consumerkey, ...)
}
\arguments{
\item{org}{Org short name}

\item{store}{store short name}

\item{auth}{location auth key}

\item{consumerkey}{location consumer key}

\item{...}{internal - advanced use only}
}
\description{
Leaflogix Get Endpoint Functions
}
\section{Functions}{
\itemize{
\item \code{get_ll_population()}: Get current inventory

\item \code{get_ll_inventory()}: Get current inventory

\item \code{get_ll_stocksnapshot()}: Get inventory snapshot

\item \code{get_ll_brands()}: Get master brands list

\item \code{get_ll_categories()}: Get master categories list

\item \code{get_ll_transactions()}: Get transactions

\item \code{get_ll_products()}: Get products

\item \code{get_ll_loyalty()}: Get loyalty

\item \code{get_ll_employees()}: Get employees

\item \code{get_ll_customers()}: Get customers

}}
