% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflogix-write.R
\name{leaflogix-write}
\alias{leaflogix-write}
\alias{wrt_ll_population}
\alias{wrt_ll_transactions}
\alias{wrt_ll_stocksnapshot}
\alias{wrt_ll_inventory}
\alias{wrt_ll_products}
\alias{wrt_ll_customers}
\alias{wrt_ll_loyalty}
\alias{wrt_ll_employees}
\alias{wrt_ll_categories}
\alias{wrt_ll_brands}
\title{Write to Database}
\usage{
wrt_ll_population(ll, org, store)

wrt_ll_transactions(ll, org, store)

wrt_ll_stocksnapshot(ll, org, store)

wrt_ll_inventory(ll, org, store)

wrt_ll_products(ll, org, store)

wrt_ll_customers(ll, org, store)

wrt_ll_loyalty(ll, org, store)

wrt_ll_employees(ll, org, store)

wrt_ll_categories(ll, org, store)

wrt_ll_brands(ll, org, store)
}
\arguments{
\item{ll}{List of tables to write to database consolidated and schema leaflogix}

\item{org}{org name for write}

\item{store}{store name for write}
}
\description{
Write to Database
}
\section{Functions}{
\itemize{
\item \code{wrt_ll_population()}: write population

\item \code{wrt_ll_transactions()}: transactions

\item \code{wrt_ll_stocksnapshot()}: stock snapshot

\item \code{wrt_ll_inventory()}: inventory

\item \code{wrt_ll_products()}: products

\item \code{wrt_ll_customers()}: customers

\item \code{wrt_ll_loyalty()}: loyalty

\item \code{wrt_ll_employees()}: employees

\item \code{wrt_ll_categories()}: categories

\item \code{wrt_ll_brands()}: brands

}}
