% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdleaflogix-package.R
\docType{package}
\name{rdleaflogix-package}
\alias{rdleaflogix}
\alias{rdleaflogix-package}
\title{rdleaflogix: Data Pipeline Execution}
\description{
This package exports pipeline functionality and contains the scripts used for remote host execution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-data-science/rdleaflogix}
  \item Report bugs at \url{https://github.com/r-data-science/rdleaflogix/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
