% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{utils-api}
\alias{utils-api}
\alias{get_pos_schema}
\title{API Resources}
\usage{
get_pos_schema(pl, name)
}
\arguments{
\item{pl}{pipeline name}

\item{name}{endpoint or resource name}
}
\description{
API Resources
}
\section{Functions}{
\itemize{
\item \code{get_pos_schema()}: Get the schema for an endpoint dataset from the api

}}
\examples{
get_pos_schema("leaflogix", "Customer")
# names can be found at api.pos.dutchie.com/swagger/index.html
get_pos_schema("leaflogix", "LoyaltySnapshot")

}
