% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-build.R
\name{utils-build}
\alias{utils-build}
\alias{set_location_uuids}
\alias{check_index}
\alias{check_schema}
\alias{set_category3}
\alias{set_brands}
\title{Tools to Faciliate the Consolidation Step}
\usage{
set_location_uuids(DT)

check_index(index)

check_schema(DT, pl, name, lmsg = NULL, include = NULL, ignore = NULL)

set_category3(DT)

set_brands(DT)
}
\arguments{
\item{DT}{internal - advanced use only}

\item{index}{internal - advanced use only}

\item{pl}{internal - advanced use only}

\item{name}{internal - advanced use only}

\item{lmsg}{internal - advanced use only}

\item{include}{internal - advanced use only}

\item{ignore}{internal - advanced use only}
}
\description{
Tools to Faciliate the Consolidation Step
}
\section{Functions}{
\itemize{
\item \code{set_location_uuids()}: internal function that appends org and store uuid to built data

\item \code{check_index()}: internal function that checks input

\item \code{check_schema()}: internal function for checking consolidated data against POS schema

\item \code{set_category3()}: Build Category3 Mappings

\item \code{set_brands()}: standardize brands

}}
