% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-creds.R
\name{utils-creds}
\alias{utils-creds}
\alias{dbOrgIndex}
\alias{dbLeaflogixCreds}
\title{Helpers to Get Credentials}
\usage{
dbOrgIndex()

dbLeaflogixCreds()
}
\description{
Helpers to Get Credentials
}
\section{Functions}{
\itemize{
\item \code{dbOrgIndex()}: get list of orgs

\item \code{dbLeaflogixCreds()}: get credentials

}}
\examples{
#active_clients <- dbOrgIndex()
#creds_leaflogix <- dbLeaflogixCreds()

}
