% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{utils-data}
\alias{utils-data}
\alias{split_by_row}
\alias{str_gender}
\title{Utilities}
\usage{
split_by_row(DT)

str_gender(vec)
}
\arguments{
\item{DT}{A data table to be split into list row-wise}

\item{vec}{vector to clean}
}
\description{
General utilities used by package functions
}
\section{Functions}{
\itemize{
\item \code{split_by_row()}: split data.table row-wise into list

\item \code{str_gender()}: clean gender column

}}
