% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get.R
\name{utils-get}
\alias{utils-get}
\alias{read_org_index}
\alias{read_brand_patterns}
\alias{read_leaflogix_creds}
\alias{get_leaflogix_index}
\alias{get_pipeline_args}
\alias{..g}
\alias{..api}
\title{Get API Keys for GET Pipelines}
\usage{
read_org_index()

read_brand_patterns()

read_leaflogix_creds()

get_leaflogix_index(org = NULL, store = NULL)

get_pipeline_args(pl = "leaflogix", org_uuids = NULL)

..g(url, ...)

..api(pl, endpt, query = NULL)
}
\arguments{
\item{org}{org short name}

\item{store}{store short name}

\item{pl}{The name of the pos pipeline (e.g. 'leaflogix')}

\item{org_uuids}{optional. If provided, pipeline arguments will be filtered on these orgs only}

\item{url}{internal - advanced use only}

\item{...}{internal - advanced use only}

\item{endpt}{internal - advanced use only}

\item{query}{internal - advanced use only}
}
\description{
Get API Keys for GET Pipelines
}
\section{Functions}{
\itemize{
\item \code{read_org_index()}: lookup of API keys configured for every location given pipeline name

\item \code{read_brand_patterns()}: lookup of API keys configured for every location given pipeline name

\item \code{read_leaflogix_creds()}: lookup index of hca orgs and stores

\item \code{get_leaflogix_index()}: Get an index table containing every org store and credentials

\item \code{get_pipeline_args()}: Get inputs needed to kick off any POS get/extract pipelines

\item \code{..g()}: internal wrapper around httr::GET

\item \code{..api()}: internal tool to build the request URL

}}
