% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-run.R
\name{utils-run}
\alias{utils-run}
\alias{get_leaflogix_args}
\alias{get_ep_names}
\alias{get_db_names}
\alias{extract_index}
\alias{.check_pl_arg}
\alias{check_inputs}
\alias{check_result}
\alias{lookupIO}
\alias{lookupIO_All}
\alias{has_rows}
\alias{lookup_pl_funs}
\title{Utilities to execute pipeline}
\usage{
get_leaflogix_args(org, store, endpt, test_run)

get_ep_names(pl = "leaflogix")

get_db_names(pl = "leaflogix")

extract_index(x, org = NULL, store = NULL)

.check_pl_arg(argll)

check_inputs(argll = NULL, test_run = FALSE, org_uuids = NULL)

check_result(res = NULL)

lookupIO(pl, tab)

lookupIO_All(pl)

has_rows(DT)

lookup_pl_funs(tab)
}
\arguments{
\item{org}{org short name}

\item{store}{store short name}

\item{endpt}{api endpoint}

\item{test_run}{internal - advanced use only}

\item{pl}{The name of the pos pipeline (e.g. 'leaflogix')}

\item{x}{extracted data as data.table or list of data.tables}

\item{argll}{internal - advanced use only}

\item{org_uuids}{internal - advanced use only}

\item{res}{internal - advanced use only}

\item{tab}{The internal name assigned to an endpoint ('transactions')}

\item{DT}{internal - advanced use only}
}
\description{
Utilities to execute pipeline
}
\section{Functions}{
\itemize{
\item \code{get_leaflogix_args()}: get arguments for leaflogix run

\item \code{get_ep_names()}: lookup configured names of the endpoints for POS and required input arg for \code{extract_index}

\item \code{get_db_names()}: lookup configured names of the DB tables

\item \code{extract_index()}: create index required by build functions given an extracted data.table and endpt

\item \code{.check_pl_arg()}: internal

\item \code{check_inputs()}: validate inputs to pipeline run 'argll

\item \code{check_result()}: check the results

\item \code{lookupIO()}: lookup data flow

\item \code{lookupIO_All()}: lookup data flow

\item \code{has_rows()}: returns FALSE if NULL or has no rows

\item \code{lookup_pl_funs()}: create index required by build functions given an extracted data.table and endpt

}}
